Move migration code to a separate file
